+++ core/src/operations/decorators/mod.rs
@@ -1,4 +1,4 @@
-use alloc::{string::ToString, vec::Vec};
+use alloc::vec::Vec;
 use core::fmt;
 
 use miden_crypto::hash::blake::Blake3_256;
@@ -8,9 +8,6 @@ use serde::{Deserialize, Serialize};
 
 mod assembly_op;
 pub use assembly_op::AssemblyOp;
-
-mod debug;
-pub use debug::DebugOptions;
 
 mod debug_var;
 pub use debug_var::{DebugVarInfo, DebugVarLocation};
@@ -31,9 +28,6 @@ use crate::mast::{DecoratedOpLink, DecoratorFingerprint};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[cfg_attr(all(feature = "arbitrary", test), miden_test_serde_macros::serde_test)]
 pub enum Decorator {
-    /// Prints out information about the state of the VM based on the specified options. This
-    /// decorator is executed only in debug mode.
-    Debug(DebugOptions),
     /// Emits a trace to the host.
     Trace(u32),
 }
@@ -41,7 +35,6 @@ pub enum Decorator {
 impl Decorator {
     pub fn fingerprint(&self) -> DecoratorFingerprint {
         match self {
-            Self::Debug(debug) => Blake3_256::hash(debug.to_string().as_bytes()),
             Self::Trace(trace) => Blake3_256::hash(&trace.to_le_bytes()),
         }
     }
@@ -56,7 +49,6 @@ impl crate::prettier::PrettyPrint for Decorator {
 impl fmt::Display for Decorator {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Debug(options) => write!(f, "debug({options})"),
             Self::Trace(trace_id) => write!(f, "trace({trace_id})"),
         }
     }