+++ processor/src/lib.rs
@@ -51,9 +51,9 @@ pub use execution_options::{ExecutionOptions, ExecutionOptionsError};
 pub use fast::{BreakReason, ExecutionOutput, FastProcessor, ResumeContext};
 pub use host::{
     BaseHost, FutureMaybeSend, Host, MastForestStore, MemMastForestStore, SyncHost,
-    debug::{DefaultDebugHandler, StdoutWriter, format_value, write_interval, write_stack},
+    debug::{DefaultTraceHandler, StdoutWriter, format_value, write_interval, write_stack},
     default::{DefaultHost, HostLibrary},
-    handlers::{DebugError, DebugHandler, TraceError},
+    handlers::{TraceError, TraceHandler},
 };
 pub use miden_core::{
     EMPTY_WORD, Felt, ONE, WORD_SIZE, Word, ZERO, crypto, field, mast, precompile,